requirements.txt
streamlit
openai
requests
PyPDF2
google.generativeai
python-dotenv
streamlit_extras
numpy
pandas